When parsing, use pattern matching instead of array length
Originally, the code copied the kotlin way of using the array length of
the split to determine what kind of command to create.
But with scala we cane use destructuring pattern matching, which allows
us to conditionally select the code path AND extract the components of
the array at the same time.

Also, this commit introduces string extension methods to correct the
incorrect java types.  Not sure why scala assumes the wrong types for
java std lib methods.  But now we can call trim and split without
polluting the types with Null everywhere.